# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0"
  },
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Get all tasks (supports filtering and search)",
        "parameters": [
          { "name": "done", "in": "query", "required": false, "schema": { "type": "boolean" }, "description": "Filter by done status" },
          { "name": "search", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Search by task title" }
        ],
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Buy milk" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request (Missing title)" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a task by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Success" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Buy milk" },
                  "done": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "Bad Request (No data provided)" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Deleted successfully" },
          "404": { "description": "Task not found" }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Get task statistics",
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "Reset tasks to default state",
        "responses": {
          "200": { "description": "Success" }
        }
      }
    }
  }
}